Move assertSourcesSinks inside helpers.js
